Add files via upload
a program that uses a loop to print the integers 1 to 100, one number on each line. When a number is divisible by 5, print "Roar" on the same line, and separate them by a space. When a number is divisible by 7, print "Lions"; and when a number is divisible by 35, print "Roar Lions!".