Game Prototype
A prototype of your unique Game that you describe in your use case.   
Your prototype can be as simple or sophisticated as you wish.  It may be fully functional or it may not do anything 
-- that is up to you.  However, it must be composed of at least two PHP pages.  
These two pages must not be accessible without logging in first -- that is, end users must login before they can test your prototype.