Tracker
The Features page has the large-scale items that are prioritized by the Product Owner. Each of the Features is broken down into manageable, ideally eight-hour, tasks. Add tasks to the tracker page to fill out the work for the sprint. Do not go over the "Hours Left in Sprint."

Finally simulate various people signing up for tasks (during the stand-up meeting) and then working on those tasks. Do this for at least half of the sprint (3/16/2021). Observe the burndown chart and make sure that it is correct. Be sure to simulate when tasks do not go according to plan.